+++ Changelog.md
@@ -14,6 +14,7 @@
 * First subtask is always checked
 * Closed tasks are mixed with open tasks
 * Start and due dates of tasks are invisible
+* Empty screen when pressing refresh
 
 ## v0.3.1
 #### fixes